Fixed issue __calc_paths and added __calc_path_preference() to the init
My proposed changes are twofold:

(1) One line we have path = []  # start a new path, meaning the end of a path is reached and the list is reset. However this only gets implemented if the if statement on 291 is true. If this if statmenent is false, the recursive call gives us subpaths, and all the current path+subpaths are added. However, path = [] is never called, even though the for loop in line 287 is about to start a new iteration (to calculate paths through a new candidate). This results in incorrect paths (I have checked). Moving path = [] outside the if statement makes sure that the path is reset before every iteration of the for loop.

2) __calc_path_preference function is defined but never ran in profile.py. Since this is a static function, it cannot be called from an instance of Profile. This results in profile. path_preference_graph to always be an empty list. We could just copy and paste the same method to our rule, but this would cause a lot of time (and would be repeated every time the method is called). It is much more efficient if this method is called upon initialization, as (I believe) originally intended.